rustc 1.67.1のclipyに対応した

+++ glisp/src/fractal/hilbert.rs
@@ -124,7 +124,7 @@ impl FractalMut for Hilbert {
     }
     fn do_demo(&mut self, c: i32) -> Result<(), Error> {
         let width = 1.0;
-        self.lgth = ((width / 2.0) as f64).powi(c);
+        self.lgth = (width / 2.0_f64).powi(c);
         let y = (width - (self.lgth * (2.0_f64).powi(c - 1))) / 3.6;
         let x = width - y;
         let oldx = x;
